% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CAsnow}
\alias{CAsnow}
\title{Gridded monthly total snowfall in Canada from 1987 to 2021.}
\format{
A list containing the location information and the observations:
\describe{
\item{locs}{A 509x2 matrix with longitude and latitude for each grid cell}
\item{n_loc}{Number of locations}
\item{Y}{A list of length 509 with each element of the list containing the
observations at a location}
}
}
\source{
\url{https://climate-change.canada.ca/climate-data/#/monthly-climate-summaries}
}
\usage{
CAsnow
}
\description{
Variables containing the monthly total snowfall (in cm) in Canada from 1987 to
2021 and the location information. The data has been gridded and information about
the grid size can be found in the paper Fast and Scalable Inference for Spatial
Extreme Value Models (arxiv: 2110.07051).
}
\keyword{datasets}
