% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_design_formula}
\alias{get_design_formula}
\title{Extract the regression matrix component of a lme style random effect term
i.e., the dots in ( ... | group). Returns a list of formulas with as many values as there
are RE terms in the formula.}
\usage{
get_design_formula(x)
}
\description{
Extract the regression matrix component of a lme style random effect term
i.e., the dots in ( ... | group). Returns a list of formulas with as many values as there
are RE terms in the formula.
}
