% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_location.R
\name{grid_location}
\alias{grid_location}
\title{Grid the locations.}
\usage{
grid_location(
  lon,
  lat,
  sp.resolution = 2,
  lon.range = range(lon),
  lat.range = range(lat)
)
}
\arguments{
\item{lon}{Numeric, \code{n} longitude values}

\item{lat}{Numeric, \code{n} latitude values}

\item{sp.resolution}{Numeric, must be a single value that indicates the minimal unit length of the grid.}

\item{lon.range}{Optional vector that indicates the range of \code{lon}. Default is \code{range(lon)}.}

\item{lat.range}{Optional vector that indicates the range of \code{lat}. Default is \code{range(lat)}.}
}
\value{
An \verb{n x 3} data frame containing three variables: \code{station_ind} corresponds to unique numbers for each grid, \code{station_lon} is the longitude of the grid, \code{station_lat} is the latitude of the grid.
}
\description{
Grid the locations.
}
\details{
The longitude and latitude of each grid is calculated by averaging the lon/lat of all locations in the same grid.
}
