% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierGev_fit.R
\name{hierGev_fit}
\alias{hierGev_fit}
\title{Fit a hierarchical Gev model.}
\usage{
hierGev_fit(y, x, theta, random, sp.thres = 0)
}
\arguments{
\item{y}{Vector of \code{n} response values.}

\item{x}{\verb{n x 2}matrix of longitude and latitude of the corresponding response values.}

\item{theta}{A list of initial parameters of. See details.}

\item{random}{A vector of parameter names that are considered as random effect. Has to be one or multiple of "a", "b", "s".}

\item{sp.thres}{Thresholding value to create sparse covariance matrix. Any distance value greater than or equal to \code{sp.thres} will be set to 0. Default is 0, which means not using sparse matrix.}
}
\value{
A list of the MLE estimates of the parameters and their corresponding standard errors.
}
\description{
Fit a hierarchical Gev model.
}
\details{
This function adopts Laplace approximation using TMB model to integrate out the random effects.
Note that \verb{b, s, sigma, ell} are estimated on the log scale.
To avoid likelihood of GEV bing 0, instead of directly using the location parameter \code{a} in estimation, its transformed version is used\preformatted{logit_tau = logit((a-s/b)/y_min)
}

where \code{y_min} is any number between \code{0} and the minimum value of \code{y}.

The log-transformed random effects are assumed to follow Gaussian processes with mean 0 and covariance matrix defined by the exponential covariance function:\preformatted{cov(i,j) = sigma*exp(-|x_i - x_j|^2/ell)
}

Therefore, when specifying the initial parameters, below are some parameters that need to specify.
Must specify: \code{logit_tau}, \code{log_b}, \code{log_s}.
Optional: \code{log_sigma_a}, \code{log_sigma_b}, \code{log_sigma_s}, \code{log_ell_a}, \code{log_ell_b}, \code{log_ell_s}.
If one of the above is considered as following GP, its corresponding \code{sigma} and \code{ell} need to be specified.
For example, if assume log_b~GP, need to specify \code{theta=c(logit_tau=1,log_b=rep(0,n),log_s=1,log_sigma_b=0,log_ell_b=0)}.
}
