% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierGev_nll.R
\name{hierGev_nll}
\alias{hierGev_nll}
\title{Calculate the negative loglikelihood of the hierarchical GEV model.}
\usage{
hierGev_nll(
  a,
  logit.tau,
  log.b,
  log.s,
  log.siga,
  log.ella,
  log.sigb,
  log.ellb,
  log.sigs,
  log.ells,
  y_min,
  y,
  dd
)
}
\arguments{
\item{a}{Vector of \code{n} location paramter}

\item{logit.tau}{Transformed vector of \code{n} location paramter by \code{logit.tau = logit((a-exp(log.s)/exp(log.b))/y_min)}, where \code{y_min} is a value between 0 and expected minimum of y.}

\item{log.b}{Log-transformed vector of \code{n} scale parameters with the constraint of \code{b>0}. Considered as a random effect.}

\item{log.s}{Log-transformed vector of \code{n} shape parameters with the constraint of \code{s>0}. Considered as a random effect.}

\item{log.siga}{The squared variance parameter (scalar) on the diagonal of the covariance matrix of \code{a ~ GP} with the constraint of \code{sigma_a > 0}.}

\item{log.ella}{The parameter (scalar) that controls the impact of the distance bewteen two locations on the covariance with the constraint of \code{ell_a > 0}.}

\item{log.sigb}{The squared variance parameter (scalar) on the diagonal of the covariance matrix of \code{logb ~ GP} with the constraint of \code{sigma_b > 0}.}

\item{log.ellb}{The parameter (scalar) that controls the impact of the distance bewteen two locations on the covariance with the constraint of \code{ell_b > 0}.}

\item{log.sigs}{The squared variance parameter (scalar) on the diagonal of the covariance matrix of \code{logs ~ GP} with the constraint of \code{sigma_s > 0}.}

\item{log.ells}{The parameter (scalar) that controls the impact of the distance bewteen two locations on the covariance with the constraint of \code{ell_s > 0}.}

\item{y_min}{The scalar used to transform location parameter into \code{logit.tau}.}

\item{y}{Vector of \code{n} responses of mws values.}

\item{dd}{A \verb{n x n} distance matrix.}
}
\value{
Scalar value of the negative loglikelihood.
}
\description{
Calculate the negative loglikelihood of the hierarchical GEV model.
}
