% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierGev_sim.R
\name{hierGev_sim}
\alias{hierGev_sim}
\title{Simulate parameters and data to test a hierarchical GEV model.}
\usage{
hierGev_sim(n.obs)
}
\arguments{
\item{n.obs}{Number of observations.}
}
\value{
A list of simulated random parameters, a dist object, and response values (MWS).
}
\description{
Simulate parameters and data to test a hierarchical GEV model.
}
\details{
Note that the purpose of this package is not to simulate data from the hierarchical model.
Instead, it just conveniently provides a set of data values to test the functions.
}
