% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierGev_tmb_nll.R
\name{hierGev_tmb_nll}
\alias{hierGev_tmb_nll}
\title{The negative loglikelihood of the hierarchical GEV model given by TMB}
\usage{
hierGev_tmb_nll(y, dd, y_min, theta, random)
}
\arguments{
\item{y}{Vector of \code{n} responses of mws values.}

\item{dd}{A \verb{n x n} distance matrix.}

\item{y_min}{The scalar used to transform location parameter into \code{logit.tau}.}

\item{theta}{A list of parameters to be evaluated at in the TMB model.}

\item{random}{A vector of parameter names that are considered as random effect. Has to be one or multiple of "a", "b", "s".}
}
\value{
Scalar value of the negative loglikelihood.
}
\description{
The negative loglikelihood of the hierarchical GEV model given by TMB
}
