% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_matern.R
\name{kernel_matern}
\alias{kernel_matern}
\title{Matern covariance function}
\usage{
kernel_matern(x, sigma, kappa, nu = 1)
}
\arguments{
\item{x}{Distance measure.}

\item{sigma}{Positive parameter. (This is in fact sigma^2)}

\item{kappa}{Positive parameter.}

\item{nu}{Range parameter default to 1.}
}
\value{
Matern covariance value.
}
\description{
Matern covariance function
}
