% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc_cov.R
\name{loc_cov}
\alias{loc_cov}
\title{The spatial exponential covariance function}
\usage{
loc_cov(dd, sigma, ell)
}
\arguments{
\item{dd}{An \verb{n x n} matrix for the distances between \code{n} locations.}

\item{sigma}{The amplitude parameter (scalar) with the constraint of \code{sigma > 0}}

\item{ell}{The smoothness parameter (scalar) with the constraint of \code{ell > 0}.}
}
\value{
An \verb{n x n} covariance matrix.
}
\description{
The spatial exponential covariance function
}
