% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot.R
\name{map_plot}
\alias{map_plot}
\title{Plot values with coordinates on a map.}
\usage{
map_plot(y, x, title, ylim, legend.shrink = 0.5, legend.horizontal = FALSE)
}
\arguments{
\item{y}{Vector of \code{n} response values.}

\item{x}{\verb{n x 2}matrix of longitude and latitude of the corresponding response values.}

\item{title}{Optional. A character string of the title of the output picture.}

\item{ylim}{The scale of \code{y} to generate colors. Default is the range of \code{y}.}

\item{legend.shrink}{Relative size of the legend. Default is 0.5.}

\item{legend.horizontal}{Whether the legend is horizontal in the plot. Default is \code{FALSE}.}
}
\value{
A plot of the generated map with different colors indicating the values of y on different locations.
}
\description{
Plot values with coordinates on a map.
}
