% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ONsnow}
\alias{ONsnow}
\title{Monthly total snowfall in Ontario, Canada from 1987 to 2021.}
\format{
A data frame with 63945 rows and 7 variables with each row corresponding to a monthly
record at a weather location:
\describe{
\item{LATITUDE}{Numeric. Latitude of the weather station}
\item{LONGITUDE}{Numeric. Longitude of the weather station}
\item{STATION_NAME}{Character. Name of the weather station}
\item{CLIMATE_IDENTIFIER}{Character. Unique id of each station}
\item{LOCAL_YEAR}{Integer from 1987 to 2021. Year of the record}
\item{LOCAL_MONTH}{Integer from 1 to 12. Month of the record}
\item{TOTAL_SNOWFALL}{Positive number. Total monthly snowfall at a station in cm}
}
}
\source{
\url{https://climate-change.canada.ca/climate-data/#/monthly-climate-summaries}
}
\usage{
ONsnow
}
\description{
A dataset containing the monthly total snowfall (in cm) in Ontario, Canada from 1987 to 2021.
}
\keyword{datasets}
