% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_sim.R
\name{param_sim}
\alias{param_sim}
\title{Draw samples of both fixed and random effects in a model with scale parameter as random effect.}
\usage{
param_sim(gev_adf, y, x, n_sim = 1000, n_random, sp.thres = 0)
}
\arguments{
\item{gev_adf}{A TMB model object that has been optimized}

\item{y}{Vector of \code{n} response values.}

\item{x}{\verb{n x 2}matrix of longitude and latitude of the corresponding response values.}

\item{n_sim}{Number of simulations. Default is 1000.}

\item{n_random}{Number of random effects}

\item{sp.thres}{Thresholding value to create sparse covariance matrix. Any distance value greater than or equal to \code{sp.thres} will be set to 0. Default is 0, which means not using sparse matrix.}
}
\value{
A list of two matrices: a matrix of draws from the approximate fixed effect posterior,
and a matrix of draws from the approximate random effect posterior. Each column correspond to one simulation.
}
\description{
Draw samples of both fixed and random effects in a model with scale parameter as random effect.
}
