% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.spatialGEVpred}
\alias{print.spatialGEVpred}
\title{Print method for spatialGEVpred}
\usage{
\method{print}{spatialGEVpred}(x, ...)
}
\arguments{
\item{...}{Additional arguments for \code{print}.}

\item{object}{Object of class \code{spatialGEVpred} returned by \code{spatialGEV_predict}.}
}
\value{
Information about the prediction.
}
\description{
Print method for spatialGEVpred
}
