% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.spatialGEVsam}
\alias{print.spatialGEVsam}
\title{Print method for spatialGEVsam}
\usage{
\method{print}{spatialGEVsam}(x, ...)
}
\arguments{
\item{x}{Object of class \code{spatialGEVsam} returned by \code{spatialGEV_sample}.}

\item{...}{Additional arguments for \code{print}.}
}
\value{
Information about the object including dimension and direction to use \code{summary} on the object.
}
\description{
Print method for spatialGEVsam
}
