% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cond_normal.R
\name{sim_cond_normal}
\alias{sim_cond_normal}
\title{Create a function to simulate from the conditional normal distribution of new data given old data}
\usage{
sim_cond_normal(joint.mean, a, X.new, X.obs, kernel = kernel_exp, sigma, ell)
}
\arguments{
\item{joint.mean}{The length `n` mean vector of the MVN distribution. By default mu1 is the first `m` elements of `joint.mean`}

\item{a}{A vector of length `n-m`, the values of mu2 to condition on}

\item{X.new}{A matrix containing the coordiantes of new locations}

\item{X.obs}{A matrix containing the coordinates of observed locations}

\item{kernel}{A function that returns a matrix containing the similarity between the two arguments}

\item{sigma}{Parameter of kernel function}

\item{ell}{Parameter of kernel function}
}
\value{
A function that takes in one argument `n` as the number of samples to draw from the condition normal distribution
of `X.new` given `X.old`: either from `rmvnorm` for MVN or `rnorm` for univariate normal. The old and new data are assumed
to follow a joint multivariate normal distribution.
}
\description{
Create a function to simulate from the conditional normal distribution of new data given old data
}
\details{
The notations are consistent to the notations on the MVN wikipedia page
}
