% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulatedData}
\alias{simulatedData}
\title{Simulated dataset 1}
\format{
A list containing the simulation parameters and simulated data on a 20x20 grid:
\describe{
\item{locs}{A 400x2 matrix. First column contains longitudes and second contains latitudes}
\item{a}{A length 400 vector. GEV location parameters}
\item{logb}{A length 400 vector. Log-transformed GEV scale parameters}
\item{logs}{A scalar. Log-transformed GEV shape parameter shared across space}
\item{y}{A length 400 list of vectors which are observations simulated at each location}
}
}
\usage{
simulatedData
}
\description{
A list of data used for package testing and demos. Both \code{a} and \code{logb} are simulated on smooth
deterministic surfaces.
}
\keyword{datasets}
