% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulatedData2}
\alias{simulatedData2}
\title{Simulated dataset 2}
\format{
A list containing the simulation parameters and simulated data on a 20x20 grid:
\describe{
\item{locs}{A 400x2 matrix. First column contains longitudes and second contains latitudes}
\item{a}{A length 400 vector. GEV location parameters}
\item{logb}{A length 400 vector. Log-transformed GEV scale parameters}
\item{logs}{A length 400 vector. Log-transformed GEV shape parameters}
\item{y}{A length 400 list of vectors which are observations simulated at each location}
}
}
\usage{
simulatedData2
}
\description{
A list of data used for package testing and demos. \code{a}, \code{logb}, \code{logs} are simulated from
respective Gaussian random fields and thus are nonsmooth.
}
\keyword{datasets}
