% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.spatialGEVfit}
\alias{summary.spatialGEVfit}
\title{Summary method for spatialGEVfit}
\usage{
\method{summary}{spatialGEVfit}(object, ...)
}
\arguments{
\item{object}{Object of class \code{spatialGEVfit} returned by \code{spatialGEV_fit}.}

\item{...}{Additional arguments for \code{summary}. Not used.}
}
\value{
Point estimates and standard errors of fixed effects, random effects,
and the return levels (if specified in \code{spatialGEV_fit()}) returned by TMB.
}
\description{
Summary method for spatialGEVfit
}
