% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_draws_coverage.R
\name{check_draws_coverage}
\alias{check_draws_coverage}
\title{Posterior draw coverage check}
\usage{
check_draws_coverage(
  draws,
  expected_covr = seq(0.1, 0.99, by = 0.01),
  ci.method = "HDI"
)
}
\arguments{
\item{draws}{An `n x k` matrix with `n` being the number of draws from the posterior predictive distribution and `k` being the number of locations}

\item{expected_covr}{A vector of length `m` containing the expected lengths of credible intervals to look at}

\item{ci.method}{Either "HDI" (default) or "ETI" for calculating the credible intervals. The former is the highest density interval and the latter is the equal tail interval.}
}
\value{
A vector of `m`. The i-th entry is the proportion of locations at which the true observation is contained 
in the credible interval given by the posterior draws, with the length of credible interval given by the i-th entry 
in `expected_covr`.
}
\description{
Posterior draw coverage check
}
