% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_exp.R
\name{kernel_exp}
\alias{kernel_exp}
\title{Exponential covariance function}
\usage{
kernel_exp(x, sigma, ell)
}
\arguments{
\item{x}{Distance measure.}

\item{sigma}{The amplitude parameter (scalar) with the constraint of \code{sigma > 0}}

\item{ell}{The smoothness parameter (scalar) with the constraint of \code{ell > 0}.}
}
\value{
Exponential covariance value.
}
\description{
Exponential covariance function
}
