% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot.R
\name{map_plot}
\alias{map_plot}
\title{Plot a the response values on a map.}
\usage{
map_plot(y, x, file = "", title, ylim, width = 800, height = 600)
}
\arguments{
\item{y}{Vector of \code{n} response values.}

\item{x}{\verb{n x 2}matrix of longitude and latitude of the corresponding response values.}

\item{file}{A character string naming a file for writing.}

\item{title}{A character string of the title of the output picture.}

\item{ylim}{The scale of \code{y} to generate colors. Default is the range of \code{y}.}

\item{width}{The width of the output file. Default is 800.}

\item{height}{The height of the output file. Default is 600.}
}
\value{
A \code{png} file of the generated map with different colors indicating the values of y on different locations.
}
\description{
Plot a the response values on a map.
}
