% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.spatialGEVfit}
\alias{print.spatialGEVfit}
\title{Print method for spatialGEVfit}
\usage{
\method{print}{spatialGEVfit}(object, ...)
}
\arguments{
\item{object}{Model object of class \code{spatialGEVfit} returned by \code{spatialGEV_fit}.}

\item{...}{More arguments for \code{print}.}
}
\value{
Information about the fitted model containing number of fixed/random effects,
fitting time, convergence information, etc.
}
\description{
Print method for spatialGEVfit
}
