% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_nll.R
\name{r_nll}
\alias{r_nll}
\title{Calculate the negative loglikelihood of the spatial GEV model.}
\usage{
r_nll(y, dd, a, log_b, s, hyperparam_a, hyperparam_b, kernel = "exp")
}
\arguments{
\item{y}{List of \code{n} locations each with \code{n_obs[i]} independent GEV realizations.}

\item{dd}{An \verb{n x n} distance matrix.}

\item{a}{Vector of \code{n} location paramter}

\item{log_b}{A numeric value or a vector of \code{n} log-transformed scale parameters if considered as a random effect.}

\item{s}{Shape parameter}

\item{hyperparam_a}{A vector of hyperparameters for a. See details.}

\item{hyperparam_b}{A vector of hyperparameters for b. Must be provided if \code{log_b} is a vector. See details.}

\item{kernel}{"exp" or "matern". Kernel function used to compute the covariance matrix for spatial random effects. Default is "exp".}
}
\value{
Scalar value of the negative loglikelihood.
}
\description{
Calculate the negative loglikelihood of the spatial GEV model.
}
\details{
This function is used to test if TMB and R output the same negative loglikelihood.
If \verb{kernel="exp}, \code{hyperparam_a/b} should be \code{c(sigma_a/b, ell_a/b)}, where \code{sigma} is the amplitude hyperparameter and \code{ell} is the smoothness hyperparameter for the exponential kernel.
If \verb{kernel="matern}, \code{hyperparam_a/b} should be \code{c(sigma_a/b, kappa_a/b)}, where \code{sigma} and \code{kappa} are hyperparameters for the Matern kernel.
If only \code{a} is a spatial random effect and \code{b} is fixed, only \code{hyperparam_a} needs to be provided.
}
