% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGEV_fit.R
\name{spatialGEV_fit}
\alias{spatialGEV_fit}
\title{Fit a hierarchical spatial Gev model.}
\usage{
spatialGEV_fit(
  y,
  X,
  random,
  init_param,
  reparam_s,
  s_prior,
  kernel = "exp",
  sp_thres = -1,
  adfun_only = FALSE,
  ignore_random = FALSE,
  silent = FALSE
)
}
\arguments{
\item{y}{List of \code{n} locations each with \code{n_obs[i]} independent GEV realizations.}

\item{X}{\verb{n x 2}matrix of longitude and latitude of the corresponding response values.}

\item{random}{Either "a" or "ab". This indicates which GEV parameters are considered as random effects.}

\item{init_param}{A list of initial parameters of. See details.}

\item{reparam_s}{A flag indicating whether the shape parameter is "zero", "unconstrained", constrained to be "negative", or constrained to be "positive". See details.}

\item{s_prior}{Optional. A length 2 vector where the first element is the mean of the normal prior on s or log(s) and the second is the standard deviation.}

\item{kernel}{Kernel function for spatial random effects covariance matrix. Can be "exp" (exponential kernel), "matern" (Matern kernel), or "spde" (Matern kernel with SPDE approximation).}

\item{sp_thres}{Optional. Thresholding value to create sparse covariance matrix. Any distance value greater than or equal to \code{sp_thres} will be set to 0. Default is -1, which means not using sparse matrix.}

\item{adfun_only}{Only output the ADfun constructed using TMB?}

\item{ignore_random}{Ignore random effect?}

\item{silent}{Do not show tracing information?}

\item{...}{Arguments to pass to \code{INLA::inla.mesh.2d()}.}
}
\value{
If \code{adfun_only=TRUE}, an list given by \code{MakeADFun()} is output.
If \code{adfun_only=FALSE}, this function outpus a list containing the following:
\itemize{
\item An adfun object
\item A fit object given by calling \code{nlminb()} on the adfun
\item An object of class \code{sdreport} from TMB which contains the point estimates, standard error, and precision matrix for the fixed and random effects
}
}
\description{
Fit a hierarchical spatial Gev model.
}
\details{
This function adopts Laplace approximation using TMB model to integrate out the random effects.

The random effects are assumed to follow Gaussian processes with mean 0 and covariance matrix defined by the chosen kernel function. E.g., using the exponential kernel function:\preformatted{cov(i,j) = sigma*exp(-|x_i - x_j|^2/ell)
}

Therefore, when specifying the initial parameters, below are some parameters that need to specify.
Must specify: \code{a}, \code{log_b}, \code{s}, \code{log_sigma_a}, \code{log_ell_a}.
Optional: \code{log_sigma_b}, \code{log_ell_b}.
If the scale parameter b is considered a random effect, its corresponding GP hyperparameters \code{log_sigma_b} and \code{log_ell_b} need to be specified.
For example, if assume both a and log_b~GP, need to specify\preformatted{init.pram=list(a=rep(1,n),log_b=rep(0,n),s=1,log_sigma_a=0,log_ell_a=0, log_sigma_b=0,log_ell_b=0).
}

The order of parameters in \code{init_param} must be: a, log_b, log_s, log_sigma_a, log_ell_a, log_sigma_b, log_ell_b.
If the Matern or SPDE kernel is used, two hyperparameters \code{sigma} and \code{kappa} are present for each spatial random effect. I.e., need to specify \code{sigma_a/b} and \code{kappa_a/b} in \code{init_param}.
If reparam_s = "negative" or "postive", the initial value of \code{s} should be that of log(|s|).
}
