% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGEV_predict.R
\name{spatialGEV_predict}
\alias{spatialGEV_predict}
\title{Draw from the posterior predictive distributions at new locations based on a fitted spatialGEV model}
\usage{
spatialGEV_predict(model, X_new, X_obs, n_draw)
}
\arguments{
\item{model}{A fitted spatial GEV model returned by \code{spatialGEV_fit}}

\item{X_new}{A \verb{n_test x 2} matrix containing the coordinates of the new locations}

\item{X_obs}{A \verb{n_obs x 2} matrix containing the coordinates of the observed locations for model fitting}

\item{n_draw}{Number of draws from the posterior predictive distribution}
}
\value{
An \verb{n_draw x n_test} matrix containing the draws from the posterior predictive distributions at \code{n_test} new locations
}
\description{
Draw from the posterior predictive distributions at new locations based on a fitted spatialGEV model
}
