% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGEV_predict.R
\name{spatialGEV_predict}
\alias{spatialGEV_predict}
\title{Draw from the posterior predictive distributions at new locations based on a fitted GEV-GP model}
\usage{
spatialGEV_predict(model, X_new, n_draw)
}
\arguments{
\item{model}{A fitted spatial GEV model object of class \code{spatialGEVfit}}

\item{X_new}{A \verb{n_test x 2} matrix containing the coordinates of the new locations}

\item{n_draw}{Number of draws from the posterior predictive distribution}
}
\value{
An object of class \code{spatialGEVpred}, which is a list of the following components:
\itemize{
\item An \verb{n_draw x n_test} matrix \code{pred_y_draws} containing the draws from the posterior predictive
distributions at \code{n_test} new locations
\item An \verb{n_loc x 2} matrix \code{X_new} containing the coordinates of the test data
\item An \verb{n_loc x 2} matrix \code{X_obs} containing the coordinates of the observed data
}
}
\description{
Draw from the posterior predictive distributions at new locations based on a fitted GEV-GP model
}
\examples{
\dontrun{
set.seed(123)
library(SpatialGEV)
a <- simulatedData$a
logb <- simulatedData$logb
logs <- simulatedData$logs
y <- simulatedData$y
locs <- simulatedData$locs
n_loc = nrow(locs)
n_test <- 20
test_ind <- sample(1:n_loc, n_test)

# Obtain coordinate matrices and data lists
locs_test <- locs[test_ind,]
y_test <- y[test_ind]
locs_train <- locs[-test_ind,]
y_train <- y[-test_ind]

# Fit the GEV-GP model to the training set
train_fit <- spatialGEV_fit(y = y_train, X = locs_train, random = "ab",
                            init_param = list(a = rep(0, n_loc-n_test), 
              				log_b = rep(0, n_loc-n_test),
					s = 0,
					log_sigma_a = 1, 
                                             log_kappa_a = -2,
					log_sigma_b = 1, 
                                             log_kappa_b = -2),
	       	       reparam_s = "positive", 
		       kernel = "matern",
		       silent = TRUE)
pred <- spatialGEV_predict(model = train_fit, X_new = locs_test, n_draw = 5000)
summary(pred)
}
}
