% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGEV_sample.R
\name{spatialGEV_sample}
\alias{spatialGEV_sample}
\title{Get posterior draws from a fitted spatial GEV model.}
\usage{
spatialGEV_sample(model, n_draw, observation = TRUE)
}
\arguments{
\item{model}{A fitted spatial GEV model returned by \code{spatialGEV_fit}}

\item{n_draw}{Number of draws from the posterior distribution}

\item{observation}{whether to draw from the posterior distribution of the GEV observation?}
}
\value{
A list of matrices containing the joint posterior draws of the parameters and optionally the GEV observations
}
\description{
Get posterior draws from a fitted spatial GEV model.
}
