% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.spatialGEVpred}
\alias{summary.spatialGEVpred}
\title{Summary method for spatialGEVpred}
\usage{
\method{summary}{spatialGEVpred}(object, ...)
}
\arguments{
\item{object}{Object of class \code{spatialGEVpred} returned by \code{spatialGEV_predict}.}

\item{...}{Additional arguments for \code{summary}.}
}
\value{
Summary statistics of the posterior predictive samples.
}
\description{
Summary method for spatialGEVpred
}
