% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.spatialGEVsam}
\alias{summary.spatialGEVsam}
\title{Summary method for spatialGEVsam}
\usage{
\method{summary}{spatialGEVsam}(object, ...)
}
\arguments{
\item{object}{Object of class \code{spatialGEVsam} returned by \code{spatialGEV_sample}.}

\item{...}{Additional arguments for \code{summary}.}
}
\value{
Summary statistics of the posterior samples.
}
\description{
Summary method for spatialGEVsam
}
